Update test.c
PyLong is 32-bit on Pynq, 64bit on Ubuntu. Just convert evertying to "long". Don't use uint32_t

+++ Pynq/cEMA/test.c
@@ -5,18 +5,18 @@
 #include "ema.h"
 
 //these values should match those in the IPython Notebook.
-const uint32_t x [] = { 2000, 2370, 1537, 2995, 1297, 3308, 1416, 3234, 1892, 2845};
-const uint32_t expect_y1 [] = {1250, 1529, 1530, 1895, 1744, 2135, 1954, 2273, 2177, 2343};
-const uint32_t expect_y3 [] = {1015, 1054, 1106, 1184, 1263, 1360, 1452, 1555, 1650, 1747};
+const long x [] = { 2000, 2370, 1537, 2995, 1297, 3308, 1416, 3234, 1892, 2845};
+const long expect_y1 [] = {1250, 1529, 1530, 1895, 1744, 2135, 1954, 2273, 2177, 2343};
+const long expect_y3 [] = {1015, 1054, 1106, 1184, 1263, 1360, 1452, 1555, 1650, 1747};
 
 int main()
 {
 
-    uint32_t size = sizeof(x) / sizeof(uint32_t);
+    long size = sizeof(x) / sizeof(long);
 
-    uint32_t * c_y1 = alloca( sizeof(x) );
-    uint32_t * c_y2 = alloca( sizeof(x) );
-    uint32_t * c_y3 = alloca( sizeof(x) );
+    long * c_y1 = alloca( sizeof(x) );
+    long * c_y2 = alloca( sizeof(x) );
+    long * c_y3 = alloca( sizeof(x) );
 
     printf("Input X: ["); 
     for (int i = 0; i < size; ++i){
